requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
yfinance>=1.2.0
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
aiosqlite==0.20.0
apscheduler==3.10.4
requests==2.32.3
jinja2==3.1.5
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
